requirements.txt
certifi==2024.2.2
charset-normalizer==2.1.1
decorator==4.4.2
idna==3.6
imageio==2.34.0
imageio-ffmpeg==0.4.9
moviepy==1.0.3
mutagen==1.47.0
numpy==1.26.4
pillow==10.2.0
proglog==0.1.10
pydub==0.25.1
python-dotenv==1.0.1
pytube==15.0.0
redis==5.0.3
requests==2.28.1
setuptools==69.1.1
six==1.16.0
spotipy==2.23.0
tqdm==4.66.2
urllib3==1.26.18
youtube-search==2.1.2
yt-dlp==2025.6.30
